ODE & Quad

+++ exercises/vec_template/main.cc
@@ -2,8 +2,8 @@
 #include "vec.h"
 
 int main() {
-    pp::vec<double> a(1, 2, 3);
-    pp::vec<double> b(4, 5, 6);
+    pp::vec<float> a(1, 2, 3);
+    pp::vec<float> b(4, 5, 6);
 
     auto c = a + b;
     std::cout << c << '\n';
